Rename 'LeadCardsResource' to 'LeadCardResource'
