+++ RentDynamics.RdClient/Resources/LeadCard/LeadCardsResourceExtensions.cs
@@ -1,11 +1,11 @@
 using JetBrains.Annotations;
 
-namespace RentDynamics.RdClient.Resources.LeadCards
+namespace RentDynamics.RdClient.Resources.LeadCard
 {
     [PublicAPI]
     public static class LeadCardsResourceExtensions
     {
-        public static LeadCardVM CreateLeadCard(this LeadCardsResource resource, int communityId, LeadCardVM request)
+        public static LeadCardVM CreateLeadCard(this LeadCardResource resource, int communityId, LeadCardVM request)
             => resource.CreateLeadCardAsync(communityId, request).GetAwaiter().GetResult();
     }
 }
