+++ RentDynamics.RdClient/Resources/LeadCard/OccupantVM.cs
@@ -1,6 +1,6 @@
 using JetBrains.Annotations;
 
-namespace RentDynamics.RdClient.Resources.LeadCards
+namespace RentDynamics.RdClient.Resources.LeadCard
 {
     [PublicAPI]
     public class OccupantVM
