+++ RentDynamics.RdClient/Resources/LeadCard/PetVM.cs
@@ -1,7 +1,7 @@
 using System;
 using JetBrains.Annotations;
 
-namespace RentDynamics.RdClient.Resources.LeadCards
+namespace RentDynamics.RdClient.Resources.LeadCard
 {
     [PublicAPI]
     public class PetVM
